refactor: split the server into modules
